Verify params specify backup and remove exes of same environment
Make sure that the versions specified by the `RdiffBackupVer` and
`RdiffBackupRemoveVer` parameters are of executables running under the
same environment (Windows or WSL), and throw otherwise.

Until now it wouldn't have worked, but parameters were not validated for
this case.

Although the script could be adapted to run one executable under Windows
and the other under WSL, it is for now an unlikely use case and would
require to complicate the script a bit by managing two separate sets of
source, destination and other variables.
For now we simply disallow it.